Moved all API related classes to their own appropriate package
